Fix test suite: 604 tests, 596 pass, 8 skip (I2P needs SAM), 0 fail
- Fix self-deadlocks: log()/escalate() called under held mutex
  (l3_stealth, l2_stealth, packet_interceptor, orchestrator)
- Fix kill-switch iptables removal to match added rule; tests never
  touch real firewall (kill switch disabled in paranoid test setup)
- nfqueue_loop: poll with timeout so stop() can join the worker
- health_monitor: sleep in slices so stop() joins promptly
- Wire SQLite3 into build (HAVE_SQLITE); guard SQLCipher-only API
- Fix probe-resist auth: generate_client_auth writes real timestamp
- Case-insensitive enum from_string + cover-mode aliases
- ProtocolMorph: count connections on select, ids from 1, restore
  on_connection_opened counter
- GA: evaluate_population works without start()
- TargetProfileDB: empty ctor per spec; BurstMorpher loads defaults
- spoofer: uint8_t printed as char in ipv4/mac generators
- license: BLAKE2b min outlen 16, fold to 8-byte key
- DoH: proper per-label hostname validation
- ForensicLogger default-enabled; DPI test port aligned to 8881
- gtest suite split (TEST vs TEST_F mix), Geneva/MSS test packets,
  I2P RequireSAM skip propagation + SO_ERROR reachability check

+++ src/core/src/spoofer.cpp
@@ -396,7 +396,9 @@ bool NetworkSpoofer::rotate_all() {
 // ==================== Generators ====================
 std::string NetworkSpoofer::generate_random_ipv4() {
     std::ostringstream oss;
-    oss << "10." << csprng_byte() << "." << csprng_byte() << "." << (1 + csprng_byte() % 254);
+    oss << "10." << static_cast<unsigned>(csprng_byte()) << "."
+        << static_cast<unsigned>(csprng_byte()) << "."
+        << (1 + static_cast<unsigned>(csprng_byte()) % 254);
     return oss.str();
 }
 
@@ -420,7 +422,7 @@ std::string NetworkSpoofer::generate_random_mac() {
     oss << std::hex << std::setfill('0');
     oss << std::setw(2) << ((csprng_byte() & 0xFC) | 0x02);
     for (int i = 0; i < 5; ++i)
-        oss << ":" << std::setw(2) << csprng_byte();
+        oss << ":" << std::setw(2) << static_cast<unsigned>(csprng_byte());
     return oss.str();
 }
